+++ openapi.yaml
@@ -181,8 +181,10 @@ components:
       properties:
         password:
           type: string
+          minLength: 1
         username:
           type: string
+          minLength: 1
       required:
       - password
       - username
@@ -198,26 +200,26 @@ components:
       properties:
         title:
           type: string
-        id:
-          type: integer
         color:
           type: string
+        id:
+          type: integer
     Event:
       type: object
       properties:
-        id:
+        categoryId:
           type: integer
-        title:
-          type: string
-          minLength: 1
         start:
           type: string
           format: date-time
+        id:
+          type: integer
         end:
           type: string
           format: date-time
-        categoryId:
-          type: integer
+        title:
+          type: string
+          minLength: 1
       required:
       - categoryId
       - end
